+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 #### Next release
 
 * Add `Drawer`
+* Add `useArraySet` hook.
 * Update design of `Spinner`.
 * Add `InputSpinner` which is used in buttons and input fields.
 * Merge the two individual theme prop for calendar and text input into one common theme prop for `DateTextInput`.